+++ HISTORY.md
@@ -1,4 +1,8 @@
 # Rocksdb Change Log
+## Additional Improvements
+### New Features
+* When user uses options.force_consistency_check in RocksDb, instead of crashing the process, we now pass the error back to the users without killing the process.
+
 ## 6.4.4 (9/17/2019)
 * Fix a bug introduced 6.3 which could cause wrong results in a corner case when prefix bloom filter is used and the iterator is reseeked.